[common] fixes coming from clang-doxy
