+++ common/include/pcl/common/impl/copy_point.hpp
@@ -35,13 +35,14 @@
  *
  */
 
-#ifndef PCL_COMMON_IMPL_COPY_POINT_HPP_
-#define PCL_COMMON_IMPL_COPY_POINT_HPP_
+#pragma once
 
 #include <pcl/point_types.h>
 #include <pcl/type_traits.h>
 #include <pcl/for_each_type.h>
 #include <pcl/common/concatenate.h>
+#include <pcl/common/copy_point.h>
+
 
 namespace pcl
 {
@@ -140,6 +141,4 @@ copyPoint (const PointInT& point_in, PointOutT& point_out)
 }
 
 } // namespace pcl
-
-#endif //PCL_COMMON_IMPL_COPY_POINT_HPP_
