+++ common/include/pcl/common/impl/intensity.hpp
@@ -38,10 +38,12 @@
  *
  */
 
-#ifndef PCL_COMMON_INTENSITY_FIELD_ACCESSOR_IMPL_HPP
-#define PCL_COMMON_INTENSITY_FIELD_ACCESSOR_IMPL_HPP
+#pragma once
 
+#include <pcl/common/intensity.h>
 #include <pcl/point_types.h>
+
+
 namespace pcl
 {
   namespace common
@@ -551,4 +553,3 @@ namespace pcl
   }
 }
 
-#endif
