+++ common/include/pcl/common/impl/intersections.hpp
@@ -35,11 +35,12 @@
  *
  */
 
-#ifndef PCL_COMMON_INTERSECTIONS_IMPL_HPP_
-#define PCL_COMMON_INTERSECTIONS_IMPL_HPP_
+#pragma once
 
+#include <pcl/common/intersections.h>
 #include <pcl/pcl_macros.h>
 #include <pcl/console/print.h>
+
 
 namespace pcl
 {
@@ -171,6 +172,4 @@ threePlanesIntersection (const Eigen::Matrix<Scalar, 4, 1> &plane_a,
 }
 
 } // namespace pcl
-
-#endif  //PCL_COMMON_INTERSECTIONS_IMPL_HPP
