+++ common/include/pcl/point_types.h
@@ -39,14 +39,8 @@
 
 #pragma once
 
-#include <pcl/PCLPointField.h>
-#include <pcl/register_point_struct.h>
-
-#include <boost/mpl/contains.hpp>
-#include <boost/mpl/fold.hpp>
-#include <boost/mpl/vector.hpp>
-
 #include <bitset>
+
 
 /**
   * \file pcl/point_types.h
@@ -349,563 +343,7 @@ namespace pcl
     * \ingroup common
     */
   struct PointDEM;
-}
-
+} // namespace pcl
 /** @} */
 
-#include <pcl/impl/point_types.hpp>  // Include struct definitions
-
-// ==============================
-// =====POINT_CLOUD_REGISTER=====
-// ==============================
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_RGB,
-    (std::uint32_t, rgba, rgba)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::RGB, pcl::_RGB)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_Intensity,
-    (float, intensity, intensity)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::Intensity, pcl::_Intensity)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_Intensity8u,
-    (std::uint8_t, intensity, intensity)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::Intensity8u, pcl::_Intensity8u)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_Intensity32u,
-    (std::uint32_t, intensity, intensity)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::Intensity32u, pcl::_Intensity32u)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointXYZ,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointXYZ, pcl::_PointXYZ)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointXYZRGBA,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (std::uint32_t, rgba, rgba)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointXYZRGBA, pcl::_PointXYZRGBA)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointXYZRGB,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, rgb, rgb)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointXYZRGB, pcl::_PointXYZRGB)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointXYZRGBL,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (std::uint32_t, rgba, rgba)
-    (std::uint32_t, label, label)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointXYZRGBL, pcl::_PointXYZRGBL)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointXYZHSV,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, h, h)
-    (float, s, s)
-    (float, v, v)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointXYZHSV, pcl::_PointXYZHSV)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointXY,
-    (float, x, x)
-    (float, y, y)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointUV,
-    (float, u, u)
-    (float, v, v)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::InterestPoint,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, strength, strength)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointXYZI,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, intensity, intensity)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointXYZI, pcl::_PointXYZI)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointXYZL,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (std::uint32_t, label, label)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Label,
-    (std::uint32_t, label, label)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_Normal,
-    (float, normal_x, normal_x)
-    (float, normal_y, normal_y)
-    (float, normal_z, normal_z)
-    (float, curvature, curvature)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::Normal, pcl::_Normal)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_Axis,
-    (float, normal_x, normal_x)
-    (float, normal_y, normal_y)
-    (float, normal_z, normal_z)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::Axis, pcl::_Axis)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointNormal,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, normal_x, normal_x)
-    (float, normal_y, normal_y)
-    (float, normal_z, normal_z)
-    (float, curvature, curvature)
-)
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointXYZRGBNormal,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, rgb, rgb)
-    (float, normal_x, normal_x)
-    (float, normal_y, normal_y)
-    (float, normal_z, normal_z)
-    (float, curvature, curvature)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointXYZRGBNormal, pcl::_PointXYZRGBNormal)
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointXYZINormal,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, intensity, intensity)
-    (float, normal_x, normal_x)
-    (float, normal_y, normal_y)
-    (float, normal_z, normal_z)
-    (float, curvature, curvature)
-)
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointXYZLNormal,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (std::uint32_t, label, label)
-    (float, normal_x, normal_x)
-    (float, normal_y, normal_y)
-    (float, normal_z, normal_z)
-    (float, curvature, curvature)
-)
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointWithRange,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, range, range)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointWithViewpoint,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, vp_x, vp_x)
-    (float, vp_y, vp_y)
-    (float, vp_z, vp_z)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointWithViewpoint, pcl::_PointWithViewpoint)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::MomentInvariants,
-    (float, j1, j1)
-    (float, j2, j2)
-    (float, j3, j3)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PrincipalRadiiRSD,
-    (float, r_min, r_min)
-    (float, r_max, r_max)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Boundary,
-    (std::uint8_t, boundary_point, boundary_point)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PrincipalCurvatures,
-    (float, principal_curvature_x, principal_curvature_x)
-    (float, principal_curvature_y, principal_curvature_y)
-    (float, principal_curvature_z, principal_curvature_z)
-    (float, pc1, pc1)
-    (float, pc2, pc2)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PFHSignature125,
-    (float[125], histogram, pfh)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PFHRGBSignature250,
-    (float[250], histogram, pfhrgb)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PPFSignature,
-    (float, f1, f1)
-    (float, f2, f2)
-    (float, f3, f3)
-    (float, f4, f4)
-    (float, alpha_m, alpha_m)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::CPPFSignature,
-    (float, f1, f1)
-    (float, f2, f2)
-    (float, f3, f3)
-    (float, f4, f4)
-    (float, f5, f5)
-    (float, f6, f6)
-    (float, f7, f7)
-    (float, f8, f8)
-    (float, f9, f9)
-    (float, f10, f10)
-    (float, alpha_m, alpha_m)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PPFRGBSignature,
-    (float, f1, f1)
-    (float, f2, f2)
-    (float, f3, f3)
-    (float, f4, f4)
-    (float, r_ratio, r_ratio)
-    (float, g_ratio, g_ratio)
-    (float, b_ratio, b_ratio)
-    (float, alpha_m, alpha_m)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::NormalBasedSignature12,
-    (float[12], values, values)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::ShapeContext1980,
-    (float[1980], descriptor, shape_context)
-    (float[9], rf, rf)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::UniqueShapeContext1960,
-    (float[1960], descriptor, shape_context)
-    (float[9], rf, rf)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::SHOT352,
-    (float[352], descriptor, shot)
-    (float[9], rf, rf)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::SHOT1344,
-    (float[1344], descriptor, shot)
-    (float[9], rf, rf)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::FPFHSignature33,
-    (float[33], histogram, fpfh)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::BRISKSignature512,
-    (float, scale, brisk_scale)
-    (float, orientation, brisk_orientation)
-    (unsigned char[64], descriptor, brisk_descriptor512)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::VFHSignature308,
-    (float[308], histogram, vfh)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::GRSDSignature21,
-    (float[21], histogram, grsd)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::ESFSignature640,
-    (float[640], histogram, esf)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(pcl::GASDSignature512,
-    (float[512], histogram, gasd)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(pcl::GASDSignature984,
-    (float[984], histogram, gasd)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(pcl::GASDSignature7992,
-    (float[7992], histogram, gasd)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::Narf36,
-    (float[36], descriptor, descriptor)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::GFPFHSignature16,
-    (float[16], histogram, gfpfh)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::IntensityGradient,
-    (float, gradient_x, gradient_x)
-    (float, gradient_y, gradient_y)
-    (float, gradient_z, gradient_z)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::PointWithScale,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, scale, scale)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(pcl::PointSurfel,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, normal_x, normal_x)
-    (float, normal_y, normal_y)
-    (float, normal_z, normal_z)
-    (std::uint32_t, rgba, rgba)
-    (float, radius, radius)
-    (float, confidence, confidence)
-    (float, curvature, curvature)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_ReferenceFrame,
-    (float[3], x_axis, x_axis)
-    (float[3], y_axis, y_axis)
-    (float[3], z_axis, z_axis)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::ReferenceFrame, pcl::_ReferenceFrame)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::_PointDEM,
-    (float, x, x)
-    (float, y, y)
-    (float, z, z)
-    (float, intensity, intensity)
-    (float, intensity_variance, intensity_variance)
-    (float, height_variance, height_variance)
-)
-POINT_CLOUD_REGISTER_POINT_WRAPPER(pcl::PointDEM, pcl::_PointDEM)
-
-namespace pcl
-{
-  // Allow float 'rgb' data to match to the newer uint32 'rgba' tag. This is so
-  // you can load old 'rgb' PCD files into e.g. a PointCloud<PointXYZRGBA>.
-  template<typename PointT>
-  struct FieldMatches<PointT, fields::rgba>
-  {
-    bool operator() (const pcl::PCLPointField& field)
-    {
-      if (field.name == "rgb")
-      {
-        // For fixing the alpha value bug #1141, the rgb field can also match
-        // uint32.
-        return ((field.datatype == pcl::PCLPointField::FLOAT32 ||
-                 field.datatype == pcl::PCLPointField::UINT32) &&
-                field.count == 1);
-      }
-      else
-      {
-        return (field.name == traits::name<PointT, fields::rgba>::value &&
-                field.datatype == traits::datatype<PointT, fields::rgba>::value &&
-                field.count == traits::datatype<PointT, fields::rgba>::size);
-      }
-    }
-  };
-  template<typename PointT>
-  struct FieldMatches<PointT, fields::rgb>
-  {
-    bool operator() (const pcl::PCLPointField& field)
-    {
-      if (field.name == "rgba")
-      {
-        return (field.datatype == pcl::PCLPointField::UINT32 &&
-                field.count == 1);
-      }
-      else
-      {
-        // For fixing the alpha value bug #1141, rgb can also match uint32
-        return (field.name == traits::name<PointT, fields::rgb>::value &&
-                (field.datatype == traits::datatype<PointT, fields::rgb>::value ||
-                 field.datatype == pcl::PCLPointField::UINT32) &&
-                field.count == traits::datatype<PointT, fields::rgb>::size);
-      }
-    }
-  };
-
-  namespace traits
-  {
-
-    /** \brief Metafunction to check if a given point type has a given field.
-     *
-     *  Example usage at run-time:
-     *
-     *  \code
-     *  bool curvature_available = pcl::traits::has_field<PointT, pcl::fields::curvature>::value;
-     *  \endcode
-     *
-     *  Example usage at compile-time:
-     *
-     *  \code
-     *  BOOST_MPL_ASSERT_MSG ((pcl::traits::has_field<PointT, pcl::fields::label>::value),
-     *                        POINT_TYPE_SHOULD_HAVE_LABEL_FIELD,
-     *                        (PointT));
-     *  \endcode
-     */
-    template <typename PointT, typename Field>
-    struct has_field : boost::mpl::contains<typename pcl::traits::fieldList<PointT>::type, Field>::type
-    { };
-
-    /** Metafunction to check if a given point type has all given fields. */
-    template <typename PointT, typename Field>
-    struct has_all_fields : boost::mpl::fold<Field,
-                                             boost::mpl::bool_<true>,
-                                             boost::mpl::and_<boost::mpl::_1,
-                                                              has_field<PointT, boost::mpl::_2> > >::type
-    { };
-
-    /** Metafunction to check if a given point type has any of the given fields. */
-    template <typename PointT, typename Field>
-    struct has_any_field : boost::mpl::fold<Field,
-                                            boost::mpl::bool_<false>,
-                                            boost::mpl::or_<boost::mpl::_1,
-                                                            has_field<PointT, boost::mpl::_2> > >::type
-    { };
-
-    /** \brief Traits defined for ease of use with fields already registered before
-     *
-     * has_<fields to be detected>: struct with `value` datamember defined at compiletime
-     * has_<fields to be detected>_v: constexpr boolean
-     * Has<Fields to be detected>: concept modelling name alias for `enable_if`
-     */
-
-    /** Metafunction to check if a given point type has x and y fields. */
-    template <typename PointT>
-    struct has_xy : has_all_fields<PointT, boost::mpl::vector<pcl::fields::x,
-                                                              pcl::fields::y> >
-    { };
-
-    template <typename PointT>
-    constexpr auto has_xy_v = has_xy<PointT>::value;
-
-    template <typename PointT>
-    using HasXY = std::enable_if_t<has_xy_v<PointT>, bool>;
-
-    template <typename PointT>
-    using HasNoXY = std::enable_if_t<!has_xy_v<PointT>, bool>;
-
-    /** Metafunction to check if a given point type has x, y, and z fields. */
-    template <typename PointT>
-    struct has_xyz : has_all_fields<PointT, boost::mpl::vector<pcl::fields::x,
-                                                               pcl::fields::y,
-                                                               pcl::fields::z> >
-    { };
-
-    template <typename PointT>
-    constexpr auto has_xyz_v = has_xyz<PointT>::value;
-
-    template <typename PointT>
-    using HasXYZ = std::enable_if_t<has_xyz_v<PointT>, bool>;
-
-    template <typename PointT>
-    using HasNoXYZ = std::enable_if_t<!has_xyz_v<PointT>, bool>;
-
-    /** Metafunction to check if a given point type has normal_x, normal_y, and
-      * normal_z fields. */
-    template <typename PointT>
-    struct has_normal : has_all_fields<PointT, boost::mpl::vector<pcl::fields::normal_x,
-                                                                  pcl::fields::normal_y,
-                                                                  pcl::fields::normal_z> >
-    { };
-
-    template <typename PointT>
-    constexpr auto has_normal_v = has_normal<PointT>::value;
-
-    template <typename PointT>
-    using HasNormal = std::enable_if_t<has_normal_v<PointT>, bool>;
-
-    template <typename PointT>
-    using HasNoNormal = std::enable_if_t<!has_normal_v<PointT>, bool>;
-
-    /** Metafunction to check if a given point type has curvature field. */
-    template <typename PointT>
-    struct has_curvature : has_field<PointT, pcl::fields::curvature>
-    { };
-
-    template <typename PointT>
-    constexpr auto has_curvature_v = has_curvature<PointT>::value;
-
-    template <typename PointT>
-    using HasCurvature = std::enable_if_t<has_curvature_v<PointT>, bool>;
-
-    template <typename PointT>
-    using HasNoCurvature = std::enable_if_t<!has_curvature_v<PointT>, bool>;
-
-    /** Metafunction to check if a given point type has intensity field. */
-    template <typename PointT>
-    struct has_intensity : has_field<PointT, pcl::fields::intensity>
-    { };
-
-    template <typename PointT>
-    constexpr auto has_intensity_v = has_intensity<PointT>::value;
-
-    template <typename PointT>
-    using HasIntensity = std::enable_if_t<has_intensity_v<PointT>, bool>;
-
-    template <typename PointT>
-    using HasNoIntensity = std::enable_if_t<!has_intensity_v<PointT>, bool>;
-
-    /** Metafunction to check if a given point type has either rgb or rgba field. */
-    template <typename PointT>
-    struct has_color : has_any_field<PointT, boost::mpl::vector<pcl::fields::rgb,
-                                                                pcl::fields::rgba> >
-    { };
-
-    template <typename PointT>
-    constexpr auto has_color_v = has_color<PointT>::value;
-
-    template <typename PointT>
-    using HasColor = std::enable_if_t<has_color_v<PointT>, bool>;
-
-    template <typename PointT>
-    using HasNoColor = std::enable_if_t<!has_color_v<PointT>, bool>;
-
-    /** Metafunction to check if a given point type has label field. */
-    template <typename PointT>
-    struct has_label : has_field<PointT, pcl::fields::label>
-    { };
-
-    template <typename PointT>
-    constexpr auto has_label_v = has_label<PointT>::value;
-
-    template <typename PointT>
-    using HasLabel = std::enable_if_t<has_label_v<PointT>, bool>;
-
-    template <typename PointT>
-    using HasNoLabel = std::enable_if_t<!has_label_v<PointT>, bool>;
-  }
-
-} // namespace pcl
-
-// Not strictly required, merely to preserve API for PCL users < 1.4
-#include <pcl/common/point_tests.h>
-
-#if defined _MSC_VER
-  #pragma warning(default: 4201)
-#endif
-
+#include <pcl/impl/point_types.hpp>
