+++ common/include/pcl/range_image/impl/range_image.hpp
@@ -36,11 +36,11 @@
  *
  */
 
-#ifndef PCL_RANGE_IMAGE_IMPL_HPP_
-#define PCL_RANGE_IMAGE_IMPL_HPP_
+#pragma once
 
-#include <pcl/pcl_macros.h>
 #include <pcl/common/distances.h>
+#include <pcl/range_image/range_image.h>
+#include <pcl/pcl_macros.h>
 
 namespace pcl
 {
@@ -1250,6 +1250,4 @@ RangeImage::integrateFarRanges (const PointCloudType& far_ranges)
   }
 }
 
-}  // namespace end
-#endif
-
+}  // namespace pcl
